Columbus

+++ src/pages/Banners/editHomeBanner.js
@@ -77,13 +77,15 @@ const EditBanner = () => {
   useEffect(() => {
     context.setProgress(20);
     fetchDataFromApi("/api/imageUpload").then((res) => {
-      res?.map((item) => {
-        item?.images?.map((img) => {
-          deleteImages(`/api/banners/deleteImage?img=${img}`).then((res) => {
-            deleteData("/api/imageUpload/deleteAllImages");
+      if (Array.isArray(res)) {
+        res?.map((item) => {
+          item?.images?.map((img) => {
+            deleteImages(`/api/banners/deleteImage?img=${img}`).then((res) => {
+              deleteData("/api/imageUpload/deleteAllImages");
+            });
           });
         });
-      });
+      }
     });
 
     fetchDataFromApi(`/api/banners/${id}`).then((res) => {
@@ -176,15 +178,19 @@ const EditBanner = () => {
           setTimeout(() => {
             setUploading(false);
             img_arr = [];
-            uniqueArray=[];
+            uniqueArray = [];
             fetchDataFromApi("/api/imageUpload").then((res) => {
+             if(Array.isArray(res)){
               res?.map((item) => {
                 item?.images?.map((img) => {
-                  deleteImages(`/api/banners/deleteImage?img=${img}`).then((res) => {
-                   // deleteData("/api/imageUpload/deleteAllImages");
-                  });
+                  deleteImages(`/api/banners/deleteImage?img=${img}`).then(
+                    (res) => {
+                      // deleteData("/api/imageUpload/deleteAllImages");
+                    }
+                  );
                 });
               });
+             }
             });
             context.setAlertBox({
               open: true,
@@ -200,18 +206,18 @@ const EditBanner = () => {
   const removeImg = async (index, imgUrl) => {
     const imgIndex = previews.indexOf(imgUrl);
 
-      deleteImages(`/api/banners/deleteImage?img=${imgUrl}`).then((res) => {
-        context.setAlertBox({
-          open: true,
-          error: false,
-          msg: "Image Deleted!",
-        });
+    deleteImages(`/api/banners/deleteImage?img=${imgUrl}`).then((res) => {
+      context.setAlertBox({
+        open: true,
+        error: false,
+        msg: "Image Deleted!",
       });
+    });
 
-      if (imgIndex > -1) {
-        // only splice array when item is found
-        previews.splice(index, 1); // 2nd parameter means remove one item only
-      }
+    if (imgIndex > -1) {
+      // only splice array when item is found
+      previews.splice(index, 1); // 2nd parameter means remove one item only
+    }
   };
 
   const handleChangeCategory = (event) => {
@@ -405,7 +411,6 @@ const EditBanner = () => {
                         <>
                           <input
                             type="file"
-                            
                             onChange={(e) =>
                               onChangeFile(e, "/api/homeBanner/upload")
                             }
