+++ src/pages/Products/editProduct.js
@@ -160,13 +160,15 @@ const EditUpload = () => {
     setCatData(context.catData);
 
     fetchDataFromApi("/api/imageUpload").then((res) => {
-      res?.map((item) => {
-        item?.images?.map((img) => {
-          deleteImages(`/api/products/deleteImage?img=${img}`).then((res) => {
-            deleteData("/api/imageUpload/deleteAllImages");
+      if (Array.isArray(res)) {
+        res?.map((item) => {
+          item?.images?.map((img) => {
+            deleteImages(`/api/products/deleteImage?img=${img}`).then((res) => {
+              deleteData("/api/imageUpload/deleteAllImages");
+            });
           });
         });
-      });
+      }
     });
 
     fetchDataFromApi(`/api/products/${id}`).then((res) => {
@@ -371,15 +373,19 @@ const EditUpload = () => {
           setTimeout(() => {
             setUploading(false);
             img_arr = [];
-            uniqueArray=[];
+            uniqueArray = [];
             fetchDataFromApi("/api/imageUpload").then((res) => {
-              res?.map((item) => {
-                item?.images?.map((img) => {
-                  deleteImages(`/api/products/deleteImage?img=${img}`).then((res) => {
-                    //deleteData("/api/imageUpload/deleteAllImages");
+              if (Array.isArray(res)) {
+                res?.map((item) => {
+                  item?.images?.map((img) => {
+                    deleteImages(`/api/products/deleteImage?img=${img}`).then(
+                      (res) => {
+                        //deleteData("/api/imageUpload/deleteAllImages");
+                      }
+                    );
                   });
                 });
-              });
+              }
             });
             context.setAlertBox({
               open: true,
